Feature.agent

+++ embedding/embedding.go
@@ -4,6 +4,6 @@ import "context"
 
 // Embedder ...
 type Embedder interface {
-	Embed(ctx context.Context, text string) ([]float64, error)
+	Embed(ctx context.Context, text string) ([]float32, error)
 	Dimension() int
 }
